+++ apps/storage/vitest.config.ts
@@ -2,6 +2,8 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
+    // Required — see the `apps/*` note in the root vitest.config.mts.
+    name: 'app:storage',
     environment: 'node',
     include: ['src/**/*.test.ts'],
   },
